requirements.txt
asgiref==3.7.2
bitcoin==1.1.39
blockcypher==1.0.93
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.2
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
greenlet==3.0.3
idna==3.6
Markdown==3.5.2
numpy==1.26.4
password-validator==1.0
pycryptodome==3.20.0
pymongo==4.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.27
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
